BIG: folder structure, moved dev server, some README
