+++ etc/tools/nodeShell.js
