+++ src/browser/css/app.styl
@@ -4,7 +4,7 @@
 @import './variables/*'
 @import './core/*'
 
-@import './../../web/**/*.styl'
+@import './../js/modules/**/*.styl'
 
 body
 	font-family 'Roboto'
