+++ src/browser/js/Router.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Route, IndexRoute } from 'react-router';
 
-import Index from './Index';
+import Index from './modules/Index';
 import Home from './modules/Home';
 import Profile from './modules/Profile';
 import Signup from './modules/Signup';
