--- src/web/.eslintrc
@@ -1,6 +1,0 @@
-{
-    "env": {
-        "browser": false,
-        "node": false
-    }
-}
